refactor: separate deck list section mapping
